// src/modules/bloggers-platform-module/comments/dto/comment-view-dto.ts
import { LikeStatus } from '../../posts/schemas/extendedLikesInfo.schema';

export class CommentViewDto {
  id: string;
  content: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  createdAt: Date;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
}